Testing and lot's of different adjustements

+++ src/main/java/com/qa/pokemons/repo/PokemonRepo.java
@@ -1,13 +1,10 @@
 package com.qa.pokemons.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
-
 import org.springframework.stereotype.Repository;
-
 import com.qa.pokemons.domain.Pokemon;
 
 @Repository
-public interface PokemonRepo extends JpaRepository<Pokemon, Integer> {
+public interface PokemonRepo extends JpaRepository<Pokemon, Long> {
 
 }
